% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chonky-package.R
\docType{package}
\name{chonky-package}
\alias{chonky}
\alias{chonky-package}
\title{chonky: Explore Chunk Inputs and Outputs}
\description{
Experimental tools to explore implicit chunk inputs and outputs
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
