% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestate-package.R
\docType{package}
\name{codestate-package}
\alias{codestate}
\alias{codestate-package}
\title{codestate: Determine How Code Depends On and Modifies State.}
\description{
Experimental tools to explore how code accesses and modifies
    state. Includes tools or
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
